Student Pages and Courses
All courses offered to computer science and software engineering masters
students have been added to courses_offered.sql. This can be copied to
mySQL to add all the courses to the course table. Each course has an ID,
name, and a dtype. The dtype will help for seeing if a student is allowed
to enroll in the course based on their major.
The student pages were just modified to show the title of each page. The
information is still missing.