Student Use Case Pages (#1)
* Title Changes

Got rid of the "Spring Security Tutorial" title and made them fit each
login.

* Student Buttons

Buttons for student use cases added. They currently do nothing.

* Button Formatting for Students

Fixed the color of the buttons.

* Enroll Page

Trying to create an enroll page for students and linking it with the
enroll button. Also commented out the create a new student.

* A View for each Student Use Case

A view page created, with a link to the student home page, for each use
case for the student.

* Student Pages and Courses

All courses offered to computer science and software engineering masters
students have been added to courses_offered.sql. This can be copied to
mySQL to add all the courses to the course table. Each course has an ID,
name, and a dtype. The dtype will help for seeing if a student is allowed
to enroll in the course based on their major.
The student pages were just modified to show the title of each page. The
information is still missing.

* Pages for Use Cases

Added pages for use cases for professor and admin

* Structure of pages + admin and prof pages

Added a structure of Student pages, such as enroll, schedule , grades and
tuition. The student info is still missing on these pages. Added pages for
the admin and professor use cases but there is only a title and a button
to get back to the home page on them.

* Course Prerequisite

Added the prerequisite courses according to the master in comp and soen
regulations. Also removed courses which were meant for phd students.

* Pages for Prof and Admin Use Cases

Added a structure to the admin and professor use case pages.