Enroll for each semester
The student can now choose between fall, winter, and summer enrollment.
Added courses for winter 2018 enrollment, still need to add more and to
add courses for summer 2018 enrollment. There is not cart showing for the
winter 2018 enrollment, and when a student enrolls in a new course, it is
not showing.
Separated some of the methods in SecurityConfiguration into CoursesOffered
to make it easier to add and modify the methods related to creating
courses.