Tuition has future and current based on enrollment
Student enrollment and admin enrollment still in progress. The student is
now enrolled in one course in winter 2018 but when they try to add a
course, it adds multiple times and does not drop. Admin still cannot
access student course history. The message was modified to be more obvious
for enrollment and tuition.